Move attention mixin into separate file
