+++ optimum/graphcore/generation/utils.py
@@ -13,11 +13,10 @@
 #  limitations under the License.
 
 import contextlib
-import copy
 import json
 import os
 import warnings
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import Any, Dict, Optional, Union
 
 import torch
 from torch import nn
@@ -51,7 +50,6 @@ from transformers.utils.versions import require_version
 logger = logging.get_logger(__name__)
 
 
-FLOAT16_LIMIT = 1e4
 MODELS_SUPPORTING_KV_CACHE = set()
 
 
@@ -72,122 +70,6 @@ def graph_profile_dir_append(append: str):
     finally:
         if poplar_engine_options_original:
             os.environ["POPLAR_ENGINE_OPTIONS"] = poplar_engine_options_original
-
-
-class IPUAttentionMixin:
-    _kv_cache_initialised: bool = False
-
-    @property
-    def kv_cache_initialised(self) -> bool:
-        return self._kv_cache_initialised
-
-    def _create_kv_cache(self, cache_shape: Tuple[int], dtype: torch.dtype, uses_beams=False):
-        self.register_buffer("_generation_step", torch.tensor([0], dtype=torch.int32), persistent=False)
-        self.register_buffer("_k_cache", torch.zeros(cache_shape, dtype=dtype), persistent=False)
-        self.register_buffer("_v_cache", torch.zeros(cache_shape, dtype=dtype), persistent=False)
-        if uses_beams:
-            self.register_buffer("_beam_idx", torch.arange(cache_shape[0], dtype=torch.int32), persistent=False)
-        self._kv_cache_initialised = True
-
-    def _delete_kv_cache(self):
-        if not self._kv_cache_initialised:
-            return
-
-        del self._generation_step
-        del self._k_cache
-        del self._v_cache
-        if hasattr(self, "_beam_idx"):
-            del self._beam_idx
-        del self._kv_cache_initialised
-
-    @classmethod
-    def from_model(
-        cls,
-        attention_layer: torch.nn.Module,
-        use_cache: bool = False,
-        batch_size: int = 1,
-        max_length: int = 128,
-        num_beams: int = 1,
-        dtype: torch.dtype = torch.float16,
-    ):
-        clone = copy.deepcopy(attention_layer)
-        clone.__class__ = cls
-
-        if use_cache:
-            clone._create_kv_cache(
-                (batch_size * num_beams, clone.num_heads, max_length, clone.head_dim),
-                dtype=dtype,
-                uses_beams=num_beams > 1,
-            )
-
-        return clone
-
-    def to_model(self, cls) -> torch.nn.Module:
-        self._delete_kv_cache()
-
-        original = copy.deepcopy(self)
-        original.__class__ = cls
-        return original
-
-    def add_to_kv_cache(self, key: torch.Tensor, value: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
-        if not self.kv_cache_initialised:
-            raise ValueError(
-                f"{self.__class__.__name__} assumes that self-attention has KV caching enabled. "
-                f"Please instantiate using `{self.__class__.__name__}.from_model()` so the KV "
-                "cache can be created."
-            )
-
-        if self.training:
-            raise RuntimeError("KV caching is currently only supported for inference.")
-
-        expected_key_shape, expected_value_shape = list(self._k_cache.shape), list(self._v_cache.shape)
-        expected_key_shape[-2] = 1
-        expected_value_shape[-2] = 1
-        if list(key.shape) != expected_key_shape:
-            raise ValueError(f"Expected key shape {expected_key_shape}, received {list(key.shape)}.")
-        if list(value.shape) != expected_value_shape:
-            raise ValueError(f"Expected value shape {expected_value_shape}, received {list(value.shape)}.")
-
-        # For now assume that generation will always start from step 0.
-        reset_kv_cache = self._generation_step == 0
-        self._k_cache *= 1 - reset_kv_cache.to(self._k_cache.dtype)
-        self._v_cache *= 1 - reset_kv_cache.to(self._v_cache.dtype)
-
-        if hasattr(self, "_beam_idx"):
-            # For beam search, permute the cache since inputs are permuted on host.
-            _k_cache = torch.index_select(self._k_cache, 0, self._beam_idx)
-            _v_cache = torch.index_select(self._v_cache, 0, self._beam_idx)
-            self._k_cache.copy_(_k_cache)
-            self._v_cache.copy_(_v_cache)
-
-        # Dynamic update leads to uneven tile placement, and scatter leads to large re-arrangements,
-        # so use a brute force matmul approach which empirically seems best for now.
-        bsz, heads, src_len, head_dim = self._k_cache.shape
-        mm_mask = (torch.arange(src_len) == self._generation_step).view(src_len, 1)
-        _key = torch.matmul(mm_mask.to(key.dtype), key.view(bsz * heads, 1, head_dim))
-        _value = torch.matmul(mm_mask.to(value.dtype), value.view(bsz * heads, 1, head_dim))
-        self._k_cache += _key.view(self._k_cache.shape)
-        self._v_cache += _value.view(self._v_cache.shape)
-
-        return self._k_cache, self._v_cache
-
-    def update_attention_mask(self, attention_mask: Optional[torch.Tensor] = None):
-        bsz, _, src_len, _ = self._k_cache.shape
-        mask = torch.full((1, src_len), -FLOAT16_LIMIT)
-        mask_cond = torch.arange(src_len).view(1, src_len)
-        mask.masked_fill_(mask_cond < self._generation_step + 1, 0)
-        mask = mask.to(self._k_cache.dtype)
-        mask = mask.expand(bsz, 1, 1, src_len)
-
-        if attention_mask is not None:
-            if attention_mask.size() != mask.size():
-                raise ValueError(
-                    f"Attention mask does not match expected KV cache mask dimensions. "
-                    f"Received: {attention_mask.size()}, expected {mask.size()}."
-                )
-            mask = mask + attention_mask
-
-        return mask
 
 
 class _IndexedInputLinear(nn.Module):
